+++ packages/ses/test/get-global-intrinsics.test.js
@@ -62,6 +62,8 @@ test.skip('getGlobalIntrinsics', () => {
     'WeakSet',
     // https://github.com/endojs/endo/issues/550
     'AggregateError',
+    // https://github.com/tc39/proposal-explicit-resource-management
+    'SuppressedError',
 
     // *** 18.4 Other Properties of the Global Object